% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_headers.R
\name{determine_headers}
\alias{determine_headers}
\title{Determine the headers of a portfolio CSV to import}
\usage{
determine_headers(filepath)
}
\arguments{
\item{filepath}{A character vector containing an absolute or relative path to
a single portfolio CSV}
}
\value{
A named character vector containing the names of the headers in the
portfolio CSV that match the proper header names expected by
pacta.portfolio.analysis. The name of each element will be the proper
column name it matches to.
}
\description{
This function will return a named vector giving the names of the headers in
the portfolio CSV that match the proper header names expected by
pacta.portfolio.analysis. The name of each element will be the proper column
name it matches to.
}
