% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv_specs.R
\name{get_csv_specs}
\alias{get_csv_specs}
\title{Get a data frame of CSV specifications}
\usage{
get_csv_specs(
  files,
  expected_colnames = c("Investor.Name", "Portfolio.Name", "ISIN", "MarketValue",
    "Currency")
)
}
\arguments{
\item{files}{A character vector containing absolute or relative paths to
portfolio CSVs, or a directory containing portfolio CSVs}

\item{expected_colnames}{A character vector containing the names of the
columns expected in the portfolio CSVs}
}
\value{
A data frame (invisibly) containing one row per portfolio CSV with
columns for each identified specification
}
\description{
This function will return a data frame with numerous specifications for every
CSV file passed in the \code{files} argument.
}
