% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_numerical_mark.R
\name{guess_numerical_mark}
\alias{guess_numerical_mark}
\title{Guess the numerical marks in the \code{market_value} column of a portfolio CSV}
\usage{
guess_numerical_mark(filepaths, type = "decimal")
}
\arguments{
\item{filepaths}{A character vector}

\item{type}{A single character string, either "decimal" or "grouping"}
}
\value{
A character vector the same length as \code{filepaths} containing a single
character string defining the guessed numerical mark for each portfolio CSV
}
\description{
This function will guess the numerical marks in the \code{market_value} column of
a portfolio CSV. It will return a single character string containing the
guessed decimal or thousands grouping mark, depending on the value passed to
\code{type}, for each portfolio CSV passed in \code{filepaths}.
}
